requirements.txt
selenium 
seleniumbase
python-dotenv
flask
webdriver_manager
mysql-connector